% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies19.R
\docType{data}
\name{constituencies19}
\alias{constituencies19}
\title{2017/19 UK constituencies}
\format{
An \code{sf} dataframe with 650 rows and 3 variables:
\describe{
\item{con_code}{Factor Constituency code}
\item{con_name}{Factor Constituency name}
\item{geometry}{sfc_GEOMETRY Constituency boundaries 2017/19}
}
}
\source{
\url{https://ckan.publishing.service.gov.uk/dataset/westminster-parliamentary-constituencies-december-2019-boundaries-uk-bfe1/resource/a13d5188-2a5d-4867-a93d-509ea4235299}
}
\usage{
constituencies19
}
\description{
Boundary data for 2017/19 UK General Election constituencies.
}
\keyword{datasets}
