% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexoutline24.R
\docType{data}
\name{hexoutline24}
\alias{hexoutline24}
\title{Boundary of UK from hexagonal constituency representation 2024}
\format{
An \code{sf} dataframe with 1 row and 2 variables:
\describe{
\item{rmapshaperid}{unknown rmapshaper-generated id column}
\item{geometry}{sfc_GEOMETRY UK outline of hexagon constituencies 2024}
}
}
\source{
\url{https://observablehq.com/@jwolondon/uk-election-2024-boundary-data}
}
\usage{
hexoutline24
}
\description{
The constituencies represented as hexagons are used to create a UK outline.
}
\keyword{datasets}
