% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexregions24.R
\docType{data}
\name{hexregions24}
\alias{hexregions24}
\title{Boundaries of regions from hexagonal constituency representation 2024}
\format{
An \code{sf} dataframe with 12 rows and 2 variables:
\describe{
\item{reg_name}{Factor Region name}
\item{geometry}{sfc_GEOMETRY Region outlines of hexagon constituencies 2024}
}
}
\source{
\url{https://observablehq.com/@jwolondon/uk-election-2024-boundary-data}
}
\usage{
hexregions24
}
\description{
The constituencies represented as hexagons are used to create region outlines.
}
\keyword{datasets}
