% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_palette.R
\docType{data}
\name{party_palette}
\alias{party_palette}
\title{Party Colour Palette}
\format{
A named character vector of length 16:
\describe{
\item{Names}{Party abbreviations (e.g., "con" for Conservative, "lab" for Labour, "ld" for Liberal Democrats, etc.).}
\item{Values}{Hexadecimal colour codes representing each party’s colour (e.g., "#0087DC" for Conservative, "#E4003B" for Labour, etc.).}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Politics_of_the_United_Kingdom/Index_of_United_Kingdom_political_parties_meta_attributes}
}
\usage{
party_palette
}
\description{
A named character vector providing hex colour codes for different UK political parties.
}
\examples{
party_palette["con"]  # Get the Conservative Party colour
party_palette["lab"]  # Get the Labour Party colour

}
\keyword{datasets}
