% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes24.R
\docType{data}
\name{votes24}
\alias{votes24}
\title{Vote count for each party in 2024}
\format{
An \code{sf} dataframe with 650 rows and 35 variables:
\describe{
\item{con_code}{Factor Constituency code}
\item{constituency_name}{Factor Constituency name}
\item{region_name}{Factor Region name}
\item{first_party24}{Factor Party with most votes in a constituency}
\item{second_party24}{Factor Party with second-most votes in a constituency}
\item{third_party24}{Factor Party with third-most votes in a constituency}
\item{con24}{int Number of votes for the Conservative Party}
\item{lab24}{int Number of votes for the Labour Party}
\item{ld24}{int Number of votes for the Liberal Democrats}
\item{ruk24}{int Number of votes for Reform UK (in other datasets \code{ruk} represents the United Kingdom Independence Party in 2017 election and the Brexit Party in 2019)}
\item{green24}{int Number of votes for the Green Party}
\item{snp24}{int Number of votes for the Scottish National Party}
\item{pc24}{int Number of votes for Plaid Cymru}
\item{dup24}{int Number of votes for the Democratic Unionist Party}
\item{sf24}{int Number of votes for Sinn Fein}
\item{sdlp24}{int Number of votes for the Social Democratic and Labour Party}
\item{uup24}{int Number of votes for the Ulster Unionist Party}
\item{apni24}{int Number of votes for the Alliance Party of Northern Ireland}
\item{other24}{int Number of votes for others}
\item{valid_votes24}{int Number of valid votes in a constituency}
\item{majority}{int Difference in vote count between first and second-placed parties}
\item{con24_pct}{num Percentage of valid votes for the Conservative Party}
\item{lab24_pct}{num Percentage of valid votes for the Labour Party}
\item{ld24_pct}{num Percentage of valid votes for the Liberal Democrats}
\item{ruk24_pct}{num Percentage of valid votes for Reform UK (in other datasets \code{ruk} represents the United Kingdom Independence Party in 2017 election and the Brexit Party in 2019)}
\item{green24_pct}{num Percentage of valid votes for the Green Party}
\item{snp24_pct}{num Percentage of valid votes for the Scottish National Party}
\item{pc24_pct}{num Percentage of valid votes for Plaid Cymru}
\item{dup24_pct}{num Percentage of valid votes for the Democratic Unionist Party}
\item{sf24_pct}{num Percentage of valid votes for Sinn Fein}
\item{sdlp24_pct}{num Percentage of valid votes for the Social Democratic and Labour Party}
\item{uup24_pct}{num Percentage of valid votes for the Ulster Unionist Party}
\item{apni24_pct}{num Percentage of valid votes for the Alliance Party of Northern Ireland}
\item{other24_pct}{num Percentage of valid votes for others}
\item{geometry}{sfc_GEOMETRY Constituency boundaries 2024}
}
}
\source{
\url{https://researchbriefings.files.parliament.uk/documents/CBP-10009/HoC-GE2024-results-by-constituency.csv} \url{https://observablehq.com/@jwolondon/uk-election-2024-boundary-data}
}
\usage{
votes24
}
\description{
The number of votes obtained by each party in each constituency in the 2024 UK General Election.
}
\keyword{datasets}
