% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constituencies24.R
\docType{data}
\name{constituencies24}
\alias{constituencies24}
\title{2024 UK constituencies}
\format{
An \code{sf} dataframe with 650 rows and 5 variables:
\describe{
\item{con_code}{Factor Constituency code}
\item{con_name}{Factor Constituency name}
\item{con_abb}{Factor Three letter constituency abbreviation}
\item{reg_name}{Factor Region name}
\item{geometry}{sfc_GEOMETRY Constituency boundaries 2024}
}
}
\source{
\url{https://observablehq.com/@jwolondon/uk-election-2024-boundary-data}
}
\usage{
constituencies24
}
\description{
Boundary data for 2024 UK General Election constituencies.
}
\keyword{datasets}
